Revert "Integrate transport service"
This reverts commit bfe6fa7d8a7249449393db31c48dbd86ae0f6be8.

+++ apps/transport/transport-common/src/main/java/transport/common/AddReservationEvent.java
@@ -9,5 +9,5 @@ import lombok.experimental.SuperBuilder;
 @SuperBuilder
 public class AddReservationEvent extends BaseEvent {
   private String userName;
-  private Long numberOfPeople;
+  private int numberOfPeople;
 }
